# api/docs/swagger.yaml
definitions:
  models.Consumer:
    properties:
      created_at:
        type: string
      custom_id:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  models.OAuth2Credential:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      consumer:
        allOf:
        - $ref: '#/definitions/models.Consumer'
        description: Relación
      consumer_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      redirect_uris:
        items:
          type: string
        type: array
    type: object
  models.OAuth2Token:
    properties:
      access_token:
        type: string
      access_token_expiration:
        type: string
      authenticated_userid:
        type: string
      created_at:
        type: integer
      credential:
        allOf:
        - $ref: '#/definitions/models.OAuth2Credential'
        description: Relación
      credential_id:
        type: string
      id:
        type: string
      refresh_token:
        type: string
      refresh_token_expiration:
        type: string
      scope:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /oauth2/introspect:
    post:
      consumes:
      - application/json
      description: Checks if a token is valid and returns token/user info
      parameters:
      - description: Token to introspect
        in: body
        name: token
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Introspect token
      tags:
      - oauth2
  /oauth2/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      - application/json
      description: Issues OAuth2 tokens (access/refresh) for a client
      parameters:
      - description: Grant type
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Client ID
        in: formData
        name: client_id
        type: string
      - description: Client Secret
        in: formData
        name: client_secret
        type: string
      - description: Authorization code
        in: formData
        name: code
        type: string
      - description: Redirect URI
        in: formData
        name: redirect_uri
        type: string
      - description: Refresh token
        in: formData
        name: refresh_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: OAuth2 token endpoint
      tags:
      - oauth2
  /oauth2/tokens:
    post:
      consumes:
      - application/json
      description: Create a new OAuth2 token with access, refresh token, and expiration
        details
      parameters:
      - description: OAuth2 Token creation payload
        in: body
        name: token
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OAuth2Token'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an OAuth2 token
      tags:
      - oauth2
  /oauth2/tokens/{token_id}:
    delete:
      description: Delete a specific OAuth2 token by ID
      parameters:
      - description: Token ID
        in: path
        name: token_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a token by ID
      tags:
      - oauth2
swagger: "2.0"